fastapiExample/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
supabase==2.21.1
python-dotenv==1.1.1
httpx==0.28.1  